+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   to get Assist sentences : a first restart to run the integration, that will copy the sentences ; and a second
   restart, that will read the sentences on start.
 - Display in dashboard if the secondary restart is needed.
+  The alert is shown when the sentences have been changed.
 - Fill the default host when creating your Recalbox, with `recalbox.local`
 - Move recalbox entities to "instances" in order to be able to store other "global" variables, like "needs_restart"
   to show info in the dashboard.